refactor some code based on reviews
